requirements.txt
Flask==3.0.0
Flask-CORS==4.0.1
openai==1.35.0
python-dotenv==1.0.1
httpx==0.26.0
Werkzeug==3.0.1